backend/requirements.txt
Flask>=3.0.0
Flask-PyMongo>=2.3.0
PyJWT>=2.8.0
bcrypt>=4.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
lxml>=5.1.0
dnspython>=2.5.0
pymongo[srv]>=4.6.0
flask-cors>=4.0.0
